+++ server/app/models/workspace.rb
@@ -27,7 +27,9 @@ class Workspace < ApplicationRecord
   has_many :alerts, dependent: :nullify
   has_many :message_feedbacks, dependent: :nullify
   has_many :chat_messages, dependent: :nullify
-  has_many :workflows, class_name: "Agents::Workflow", dependent: :nullify
+  has_many :workflows, class_name: "Agents::Workflow", dependent: :destroy
+  has_many :workflow_runs, class_name: "Agents::WorkflowRun", dependent: :destroy
+  has_many :workflow_logs, class_name: "Agents::WorkflowLog", dependent: :nullify
 
   belongs_to :organization
   has_many :sso_configurations, through: :organization